refactor: centralize cardapio helpers
